move parts of the http controller to the core lib
