+++ canisters/control-panel/api/src/lib.rs
@@ -3,7 +3,6 @@
 
 /// User DTOs.
 mod user;
-use candid::{CandidType, Deserialize};
 pub use user::*;
 
 /// User Wallet DTOs.
@@ -17,25 +16,3 @@ pub use manage_user::*;
 /// Canister hooks DTOs.
 mod canister;
 pub use canister::*;
-
-// Http Interface (for metrics)
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct HeaderField(pub String, pub String);
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct HttpRequest {
-    pub method: String,
-    pub url: String,
-    pub headers: Vec<HeaderField>,
-    #[serde(with = "serde_bytes")]
-    pub body: Vec<u8>,
-}
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct HttpResponse {
-    pub status_code: u16,
-    pub headers: Vec<HeaderField>,
-    #[serde(with = "serde_bytes")]
-    pub body: Vec<u8>,
-}
