--- canisters/wallet/api/src/http.rs
@@ -1,22 +1,0 @@
-//! HTTP interface types for the canister.
-use candid::{CandidType, Deserialize};
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct HeaderField(pub String, pub String);
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct HttpRequest {
-    pub method: String,
-    pub url: String,
-    pub headers: Vec<HeaderField>,
-    #[serde(with = "serde_bytes")]
-    pub body: Vec<u8>,
-}
-
-#[derive(Clone, Debug, CandidType, Deserialize)]
-pub struct HttpResponse {
-    pub status_code: u16,
-    pub headers: Vec<HeaderField>,
-    #[serde(with = "serde_bytes")]
-    pub body: Vec<u8>,
-}
