+++ canisters/wallet/api/src/lib.rs
@@ -1,9 +1,6 @@
 //! Transport layer that defines the data transfer objects (DTOs) that are used to communicate
 //! with the clients of the wallet canister.
 
-mod http;
-pub use http::*;
-
 mod capabilities;
 pub use capabilities::*;
