+++ libs/ic-canister-core/Cargo.toml
@@ -21,6 +21,7 @@ prometheus = { workspace = true }
 rand_chacha = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+serde_bytes = { workspace = true }
 time = { workspace = true, features = ["formatting", "parsing"] }
 thiserror = { workspace = true }
 uuid = { workspace = true, features = ["serde", "v4"] }
