+++ libs/ic-canister-core/src/api.rs
@@ -89,3 +89,24 @@ pub type ApiResult<Data = ()> = Result<Data, ApiError>;
 
 pub type ServiceResult<CompletedOperation = (), FailedOperation = ApiError> =
     Result<CompletedOperation, FailedOperation>;
+
+// Http Interface
+#[derive(Clone, Debug, CandidType, Deserialize)]
+pub struct HeaderField(pub String, pub String);
+
+#[derive(Clone, Debug, CandidType, Deserialize)]
+pub struct HttpRequest {
+    pub method: String,
+    pub url: String,
+    pub headers: Vec<HeaderField>,
+    #[serde(with = "serde_bytes")]
+    pub body: Vec<u8>,
+}
+
+#[derive(Clone, Debug, CandidType, Deserialize)]
+pub struct HttpResponse {
+    pub status_code: u16,
+    pub headers: Vec<HeaderField>,
+    #[serde(with = "serde_bytes")]
+    pub body: Vec<u8>,
+}
